Updates for DevPreview: - remove Context.payloadVersion - rename package.json description - currently removing references to `cloudevents` in dependencies, will pull in 1.0.0 with types in later work as needed - move source from lib/ to src/ with corresponding updates to tests
