+++ test/unit/sobjects/SObjectTests.ts
@@ -2,7 +2,7 @@
 import { assert, expect } from 'chai';
 import 'mocha';
 
-import { SObject } from '../../../lib';
+import { SObject } from '../../../src';
 
 describe('SObject Tests', () => {
     let sObject: SObject = null;
